feat: make install pwa first step on the setup flow
