+++ src/app/(setup)/setup/page.tsx
@@ -13,6 +13,7 @@ export default function SetupPage() {
     const [currentStepIndex, setCurrentStepIndex] = useState(0)
     const [deferredPrompt, setDeferredPrompt] = useState<BeforeInstallPromptEvent | null>(null)
     const [canInstall, setCanInstall] = useState(false)
+    const [deviceType, setDeviceType] = useState<'ios' | 'android' | 'desktop'>('desktop')
 
     useEffect(() => {
         // Store the install prompt
@@ -22,6 +23,24 @@ export default function SetupPage() {
             setDeferredPrompt(e as BeforeInstallPromptEvent)
             setCanInstall(true)
         }
+
+        // Detect device type
+        const isIOSDevice = /iPad|iPhone|iPod|Mac|Macintosh/.test(navigator.userAgent)
+        const isMobileDevice = /Android|webOS|iPad|iPhone|iPod|BlackBerry|IEMobile|Opera Mini/i.test(
+            navigator.userAgent
+        )
+
+        // For desktop, default to iOS if on Mac, otherwise Android
+        if (!isMobileDevice) {
+            setDeviceType('desktop')
+        } else {
+            if (isIOSDevice) {
+                setDeviceType('ios')
+            } else {
+                setDeviceType('android')
+            }
+        }
+
         window.addEventListener('beforeinstallprompt', handleBeforeInstallPrompt)
         return () => {
             window.removeEventListener('beforeinstallprompt', handleBeforeInstallPrompt)
@@ -56,6 +75,7 @@ export default function SetupPage() {
             direction={direction}
             deferredPrompt={deferredPrompt}
             canInstall={canInstall}
+            deviceType={deviceType}
         >
             <step.component />
         </SetupWrapper>
